pyproject.toml
[project]
name = "adriabox"
version = "0.1.0"
description = ""
authors = [
    {name = "Wearemassive78",email = "massimiliano.pupillo@studio.unibo.it"}
]
readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)"
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.poetry]
packages = [
    { include = "client", from = "src" },
    { include = "common", from = "src" },
    { include = "metadata_server", from = "src" },
    { include = "storage_node", from = "src" }
]
